pyproject.toml
[project]
name = "lifeos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dateparser>=1.2.2",
    "fastapi>=0.122.0",
    "openai>=2.8.1",
    "openai-agents>=0.6.1",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.27",
    "truststore>=0.10.4",
    "uvicorn>=0.38.0",
]
